Compile fix for GCC
<==== Mercurial data ====>
Branch: vms_4.0_gui

+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/system_administration/models/time_synchronization_servers_model.cpp
@@ -187,9 +187,9 @@ QVariant TimeSynchronizationServersModel::data(const QModelIndex& index, int rol
             switch (column)
             {
                 case OsTimeColumn:
-                    return server.osTimeOffset.count();
+                    return (qint64) server.osTimeOffset.count();
                 case VmsTimeColumn:
-                    return server.vmsTimeOffset.count();
+                    return (qint64) server.vmsTimeOffset.count();
 
                 default:
                     break;
